src/test/java/com/springbasics/springin5steps/cdi/SomeCdiBusinessTest.java
package com.springbasics.springin5steps.cdi;

import static org.junit.Assert.assertEquals;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringRunner;

import com.springbasics.springin5steps.SpringIn5StepsBasicApplication;

@RunWith(SpringRunner.class)
@ContextConfiguration(classes = SpringIn5StepsBasicApplication.class)
public class SomeCdiBusinessTest {

	@Autowired
	SomeCdiBusiness business;

	@Test
	public void testBasicScenario() {

		int actualResult = business.findGreatest();

		assertEquals(actualResult, 100);
	}

}
